Moves flags.go to types and removes utils (#385)
