+++ client/BUILD.bazel
@@ -9,7 +9,7 @@ go_library(
     visibility = ["//client:__subpackages__"],
     deps = [
         "//client/node:go_default_library",
-        "//client/utils:go_default_library",
+        "//client/types:go_default_library",
         "//shared/cmd:go_default_library",
         "//shared/debug:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
@@ -28,7 +28,7 @@ go_image(
     visibility = ["//visibility:private"],
     deps = [
         "//client/node:go_default_library",
-        "//client/utils:go_default_library",
+        "//client/types:go_default_library",
         "//shared/cmd:go_default_library",
         "//shared/debug:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
