+++ client/node/BUILD.bazel
@@ -14,7 +14,7 @@ go_library(
         "//client/rpcclient:go_default_library",
         "//client/syncer:go_default_library",
         "//client/txpool:go_default_library",
-        "//client/utils:go_default_library",
+        "//client/types:go_default_library",
         "//shared:go_default_library",
         "//shared/cmd:go_default_library",
         "//shared/database:go_default_library",
