+++ client/types/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "collation.go",
+        "flags.go",
         "interfaces.go",
         "shard.go",
     ],
@@ -19,6 +20,7 @@ go_library(
         "@com_github_ethereum_go_ethereum//rlp:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_syndtr_goleveldb//leveldb/errors:go_default_library",
+        "@com_github_urfave_cli//:go_default_library",
     ],
 )
