+++ client/types/flags.go
@@ -1,4 +1,4 @@
-package utils
+package types
 
 import (
 	"math/big"
