--- client/utils/BUILD.bazel
@@ -1,12 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["flags.go"],
-    importpath = "github.com/prysmaticlabs/prysm/client/utils",
-    visibility = ["//client:__subpackages__"],
-    deps = [
-        "//client/params:go_default_library",
-        "@com_github_urfave_cli//:go_default_library",
-    ],
-)
